% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{dbt_001_colors}
\alias{dbt_001_colors}
\title{Theme 'dbt_001': colors}
\usage{
dbt_001_colors(name = character(), reverse = FALSE, flip = FALSE, df = FALSE)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Theme 'dbt_001': colors
}
\examples{
dbt_001_colors()
}
